Se cambian nombres de las clases fragment a inglés, se usan dimens, se elimina código innecesario
