+++ .idea/misc.xml
@@ -45,10 +45,13 @@
         <entry key="..\:/Users/root/Documents/ProjectsAndroidStudio/AppPractica/app/src/main/res/layout/activity_singup.xml" value="0.18" />
         <entry key="..\:/Users/root/Documents/ProjectsAndroidStudio/AppPractica/app/src/main/res/layout/best_movies.xml" value="0.17552083333333332" />
         <entry key="..\:/Users/root/Documents/ProjectsAndroidStudio/AppPractica/app/src/main/res/layout/fragment_buscar.xml" value="0.18" />
+        <entry key="..\:/Users/root/Documents/ProjectsAndroidStudio/AppPractica/app/src/main/res/layout/fragment_home.xml" value="0.1" />
         <entry key="..\:/Users/root/Documents/ProjectsAndroidStudio/AppPractica/app/src/main/res/layout/fragment_inicio.xml" value="0.25" />
         <entry key="..\:/Users/root/Documents/ProjectsAndroidStudio/AppPractica/app/src/main/res/layout/fragment_movie_details.xml" value="0.33" />
         <entry key="..\:/Users/root/Documents/ProjectsAndroidStudio/AppPractica/app/src/main/res/layout/fragment_perfil.xml" value="0.25" />
         <entry key="..\:/Users/root/Documents/ProjectsAndroidStudio/AppPractica/app/src/main/res/layout/fragment_play.xml" value="0.17552083333333332" />
+        <entry key="..\:/Users/root/Documents/ProjectsAndroidStudio/AppPractica/app/src/main/res/layout/fragment_profile.xml" value="0.1" />
+        <entry key="..\:/Users/root/Documents/ProjectsAndroidStudio/AppPractica/app/src/main/res/layout/fragment_search.xml" value="0.1" />
         <entry key="..\:/Users/root/Documents/ProjectsAndroidStudio/AppPractica/app/src/main/res/layout/item_horizontal.xml" value="0.25" />
         <entry key="..\:/Users/root/Documents/ProjectsAndroidStudio/AppPractica/app/src/main/res/layout/item_operaciones_horizontal.xml" value="0.22" />
         <entry key="..\:/Users/root/Documents/ProjectsAndroidStudio/AppPractica/app/src/main/res/layout/layout_item_view.xml" value="0.25" />
