+++ app/src/main/res/layout/fragment_home.xml
@@ -5,7 +5,7 @@
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:background="@color/color_fondo_layout"
-    tools:context=".Fragments.InicioFragment">
+    tools:context=".Fragments.HomeFragment">
 
     <ScrollView
         android:layout_width="match_parent"
