+++ app/src/main/res/layout/fragment_play.xml
@@ -5,7 +5,6 @@
     android:layout_height="match_parent"
     tools:context=".Fragments.PlayFragment">
 
-    <!-- TODO: Update blank fragment layout -->
     <TextView
         android:gravity="center"
         android:layout_width="match_parent"
