+++ app/src/main/res/layout/fragment_profile.xml
@@ -11,7 +11,7 @@
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:layout_marginTop="25dp"
-        tools:context=".Fragments.PerfilFragment">
+        tools:context=".Fragments.ProfileFragment">
 
         <LinearLayout
             android:id="@+id/DatosPersonales"
