+++ app/src/main/res/layout/fragment_search.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".Fragments.BuscarFragment">
+    tools:context=".Fragments.SearchFragment">
 
     <LinearLayout
         android:layout_width="match_parent"
@@ -19,13 +19,13 @@
                 android:id="@+id/searchView"
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
-                app:queryHint="Buscar"
-                app:queryBackground="@color/color_transparente"
-                app:iconifiedByDefault="false"
                 android:layout_marginHorizontal="16dp"
                 android:layout_marginTop="30dp"
                 android:layout_marginBottom="16dp"
-                android:background="@drawable/searchview_background">
+                android:background="@drawable/searchview_background"
+                app:iconifiedByDefault="false"
+                app:queryBackground="@color/color_transparente"
+                app:queryHint="Buscar">
 
             </androidx.appcompat.widget.SearchView>
