+++ app/src/main/res/navigation/nav_graph.xml
@@ -3,26 +3,26 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/nav_graph"
-    app:startDestination="@id/inicioFragment">
+    app:startDestination="@id/homeFragment">
 
     <fragment
-        android:id="@+id/inicioFragment"
-        android:name="com.example.apppractica.Fragments.InicioFragment"
-        android:label="fragment_inicio"
-        tools:layout="@layout/fragment_inicio" />
+        android:id="@+id/homeFragment"
+        android:name="com.example.apppractica.Fragments.HomeFragment"
+        android:label="fragment_home"
+        tools:layout="@layout/fragment_home" />
     <fragment
-        android:id="@+id/buscarFragment"
-        android:name="com.example.apppractica.Fragments.BuscarFragment"
-        android:label="fragment_buscar"
-        tools:layout="@layout/fragment_buscar" />
+        android:id="@+id/searchFragment"
+        android:name="com.example.apppractica.Fragments.SearchFragment"
+        android:label="fragment_search"
+        tools:layout="@layout/fragment_search" />
     <fragment
         android:id="@+id/playFragment"
         android:name="com.example.apppractica.Fragments.PlayFragment"
         android:label="fragment_play"
         tools:layout="@layout/fragment_play" />
     <fragment
-        android:id="@+id/perfilFragment"
-        android:name="com.example.apppractica.Fragments.PerfilFragment"
+        android:id="@+id/profileFragment"
+        android:name="com.example.apppractica.Fragments.ProfileFragment"
         android:label="fragment_perfil"
-        tools:layout="@layout/fragment_perfil" />
+        tools:layout="@layout/fragment_profile" />
 </navigation>
